cloudrun_api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
pydantic
bittensor>=6.0.0